0.8.7v - 학생 비밀번호 변경 추가

+++ src/pages/teacher/TNotice/TNotice.tsx
@@ -30,6 +30,7 @@ export interface INotice {
 }
 
 const TNotice = () => {
+    const [handler, setHandler] = useState(false);
     const navigate = useNavigate();
     const [searchParams, setSearchParams] = useSearchParams();
     const page = searchParams.get("page");
@@ -73,16 +74,19 @@ const TNotice = () => {
 
     useEffect(() => {
         fetchData();
-    }, [page, keyword]);
+    }, [page, keyword, handler]);
 
     const deleteNotice = async () => {
-        try {
-            const response = await axios.delete(
-                `http://192.168.0.62:9988/api/notice/${noticeNos}/${teacherNo}`,
-            );
-            alert(response.data.message);
-        } catch (error) {
-            console.error(error);
+        if (window.confirm("정말 삭제하시겠습니까?")) {
+            try {
+                const response = await axios.delete(
+                    `http://192.168.0.62:9988/api/notice/${noticeNos}/${teacherNo}`,
+                );
+                alert(response.data.message);
+                setHandler(!handler);
+            } catch (error) {
+                console.error(error);
+            }
         }
     };
